Add solc compile option for test contracts

+++ src/tests/build.sh
@@ -1,26 +1,30 @@
 #!/bin/bash
 
-truffle compile || exit 1
+if [[ $1 == "solc" ]]
+then
+    mkdir -p build
+    solc -o build --bin --abi --overwrite ./contracts/*.sol
+else
+    contracts=(
+        "SolTests"
+        "SubContract"
+        "Create2Factory"
+        "SelfDestruct"
+        "ConstructorRevert"
+    )
 
-contracts=(
-    "SolTests"
-    "SubContract"
-    "Create2Factory"
-    "SelfDestruct"
-    "ConstructorRevert"
-)
+    truffle compile || exit 1
+    for contractName in "${contracts[@]}"
+        do
+            cat build/contracts/"$contractName".json | \
+              jq .bytecode | \
+              awk '{ print substr($1,4,length($1)-4) }' | \
+              tr -d '\n' \
+              > build/"$contractName".bin
 
-for contractName in "${contracts[@]}"
-    do
-        cat build/contracts/"$contractName".json | \
-          jq .bytecode | \
-          awk '{ print substr($1,4,length($1)-4) }' | \
-          tr -d '\n' \
-          > build/"$contractName".bin
-
-        cat build/contracts/"$contractName".json | \
-          jq .abi \
-          > build/"$contractName".abi
-    done
-
-rm -rf build/contracts
+            cat build/contracts/"$contractName".json | \
+              jq .abi \
+              > build/"$contractName".abi
+        done
+    rm -rf build/contracts
+fi
